Got window

+++ src/Grid.cpp
@@ -6,9 +6,9 @@ Grid::Grid() : w(0) , h(0), cells(nullptr)
 
 Grid::Grid(int w, int h) : w(w) , h(h) , cells(new Cell[w*h])
 {
-    for(size_t j = 0; j < h; j++)
+    for(int j = 0; j < h; j++)
     {
-        for(size_t i = 0; i < w; i++)
+        for(int i = 0; i < w; i++)
         {
             this->cells[i + (j*this->w)] = Cell(i,j);
         }
